# docs/swagger.yaml
basePath: /
definitions:
  district-friends_internal_models.ChatRoom:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  district-friends_internal_models.ChatRoomMember:
    properties:
      id:
        type: integer
      joined_at:
        type: string
      room_id:
        type: integer
      user_id:
        type: integer
    type: object
  district-friends_internal_models.ExternalEvent:
    properties:
      date:
        description: String format for simplicity
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      location:
        type: string
      price_max:
        type: number
      price_min:
        type: number
      title:
        type: string
      type:
        description: '"movie" or "concert"'
        type: string
      url:
        type: string
    type: object
  district-friends_internal_models.Friendship:
    properties:
      created_at:
        type: string
      friend_id:
        type: integer
      id:
        type: integer
      status:
        $ref: '#/definitions/district-friends_internal_models.FriendshipStatus'
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  district-friends_internal_models.FriendshipStatus:
    enum:
    - pending
    - accepted
    - declined
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusAccepted
    - StatusDeclined
  district-friends_internal_models.Message:
    properties:
      content:
        type: string
      created_at:
        type: string
      external_event_id:
        description: String ID from external API (TMDB/Ticketmaster)
        type: string
      external_event_type:
        description: '"movie" or "concert"'
        type: string
      id:
        type: integer
      room_id:
        type: integer
      sender_id:
        type: integer
    type: object
  district-friends_internal_models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  internal_api_handlers.CreateRoomReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  internal_api_handlers.CreateUserRequest:
    properties:
      email:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  internal_api_handlers.FriendRequestReq:
    properties:
      friend_id:
        type: integer
      user_id:
        type: integer
    required:
    - friend_id
    - user_id
    type: object
  internal_api_handlers.JoinRoomReq:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
  internal_api_handlers.ShareEventReq:
    properties:
      external_event_id:
        type: string
      external_event_type:
        type: string
      user_id:
        type: integer
    required:
    - external_event_id
    - external_event_type
    - user_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a friends feature extension backend.
  title: Friends District API
  version: "1.0"
paths:
  /api/v1/events:
    get:
      description: Get a list of mock events (movies, concerts, dining)
      parameters:
      - description: Filter by event type (e.g. Movie, Concert)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/district-friends_internal_models.ExternalEvent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List Events
      tags:
      - events
  /api/v1/events/{id}:
    get:
      description: Get details for a specific event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/district-friends_internal_models.ExternalEvent'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get Event Details
      tags:
      - events
  /api/v1/friends:
    get:
      description: Retrieve a list of accepted friends for a user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/district-friends_internal_models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List friends
      tags:
      - friends
  /api/v1/friends/accept:
    post:
      consumes:
      - application/json
      description: Accept a pending friend request
      parameters:
      - description: Friend Request Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.FriendRequestReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/district-friends_internal_models.Friendship'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Accept a friend request
      tags:
      - friends
  /api/v1/friends/request:
    post:
      consumes:
      - application/json
      description: Send a friend request to another user
      parameters:
      - description: Friend Request Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.FriendRequestReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/district-friends_internal_models.Friendship'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send a friend request
      tags:
      - friends
  /api/v1/rooms:
    post:
      consumes:
      - application/json
      description: Create a new chat room
      parameters:
      - description: Room Name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CreateRoomReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/district-friends_internal_models.ChatRoom'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a chat room
      tags:
      - chat
  /api/v1/rooms/{id}/join:
    post:
      consumes:
      - application/json
      description: Add a user to a chat room
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.JoinRoomReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/district-friends_internal_models.ChatRoomMember'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Join a chat room
      tags:
      - chat
  /api/v1/rooms/{id}/messages:
    get:
      description: Retrieve message history for a chat room
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/district-friends_internal_models.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get room messages
      tags:
      - chat
  /api/v1/rooms/{id}/share:
    post:
      consumes:
      - application/json
      description: Share a specific TMDB or Ticketmaster event to a chat room
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID and External Event Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.ShareEventReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/district-friends_internal_models.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Share an external event in a chat room
      tags:
      - chat
  /api/v1/rooms/{id}/ws:
    get:
      description: Connect to real-time chat via WebSocket
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      responses: {}
      summary: WebSocket for Chat Room
      tags:
      - chat
  /api/v1/users:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/district-friends_internal_models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a user
      tags:
      - users
swagger: "2.0"