# internal/api/handlers/event_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"

	"district-friends/internal/models"
	"district-friends/internal/services"
)

type EventHandler struct {
	TMDB         *services.TMDBService
	Ticketmaster *services.TicketmasterService
}

func NewEventHandler() *EventHandler {
	return &EventHandler{
		TMDB:         services.NewTMDBService(),
		Ticketmaster: services.NewTicketmasterService(),
	}
}

// @Summary List Events
// @Description Get a list of mock events (movies, concerts, dining)
// @Tags events
// @Produce json
// @Param type query string false "Filter by event type (e.g. Movie, Concert)"
// @Success 200 {array} models.ExternalEvent
// @Failure 500 {object} map[string]interface{}
// @Router /api/v1/events [get]
func (h *EventHandler) ListEvents(c *gin.Context) {
	var allEvents []models.ExternalEvent

	eventType := c.Query("type")

	// Fetch Movies
	if eventType == "" || eventType == "movie" {
		movies, err := h.TMDB.FetchMovies()
		if err == nil {
			allEvents = append(allEvents, movies...)
		}
	}

	// Fetch Concerts
	if eventType == "" || eventType == "concert" {
		concerts, err := h.Ticketmaster.FetchConcerts()
		if err == nil {
			allEvents = append(allEvents, concerts...)
		}
	}

	c.JSON(http.StatusOK, allEvents)
}

// @Summary Get Event Details
// @Description Get details for a specific event
// @Tags events
// @Produce json
// @Param id path int true "Event ID"
// @Success 200 {object} models.ExternalEvent
// @Failure 400 {object} map[string]interface{}
// @Failure 404 {object} map[string]interface{}
// @Router /api/v1/events/{id} [get]
func (h *EventHandler) GetEvent(c *gin.Context) {
	// For this prototype, since we merge data, fetching a single event by ID across 2 different APIs 
	// would require knowing which API the ID belongs to. The frontend usually passes the type, or we search both.
	// We'll leave a simple stub here that tells the client to use the list endpoint, or we can implement a search.
	
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Single event lookup is not fully implemented for merged external APIs. Use the list endpoint."})
}